Move json schema utils to utils package
